Tic Tac Toe LLD started (made LLD using Top down approach)

+++ demo/src/main/java/com/example/demo/Blinkit/Strategy/RainSurgePricing.java
@@ -3,8 +3,8 @@ package com.example.demo.Blinkit.Strategy;
 import com.example.demo.Blinkit.Entity.Cart;
 
 public class RainSurgePricing implements PricingStrategy{
-    int rainSurgeMultiplier;
-    public RainSurgePricing(int rainSurgeMultiplier){
+    double rainSurgeMultiplier;
+    public RainSurgePricing(double rainSurgeMultiplier){
         this.rainSurgeMultiplier=rainSurgeMultiplier;
     }
     public double calculatePrice(Cart cart){
